backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
greenlet==3.1.1
edge-tts==6.1.12
anthropic==0.34.2
openai==1.51.0
python-dotenv==1.0.1
qrcode[pil]==7.4.2
Pillow==10.4.0
python-multipart==0.0.9
httpx==0.27.2
pydantic==2.9.2
